factored out services
These changes are heavily inspired by CAPA. They should improve the structure
of the code for the upcoming addition of multi-node control plane
support.
