+++ pkg/cloud/openstack/contants/constants.go
@@ -1,0 +1,7 @@
+package constants
+
+
+const (
+	OpenstackIPAnnotationKey = "openstack-ip-address"
+	OpenstackIdAnnotationKey = "openstack-resourceId"
+)
