+++ pkg/cloud/openstack/services/clouds.yaml
