+++ pkg/cloud/openstack/services/userdata/machinescript_test.go
@@ -1,4 +1,4 @@
-package machine
+package userdata
 
 import (
 	"testing"
