+++ pkg/deployer/deployer.go
@@ -1,0 +1,65 @@
+package deployer
+
+import (
+	"fmt"
+	"github.com/pkg/errors"
+	"k8s.io/klog"
+	"os"
+	providerv1 "sigs.k8s.io/cluster-api-provider-openstack/pkg/apis/openstackproviderconfig/v1alpha1"
+	constants "sigs.k8s.io/cluster-api-provider-openstack/pkg/cloud/openstack/contants"
+	clusterv1 "sigs.k8s.io/cluster-api/pkg/apis/cluster/v1alpha1"
+	"sigs.k8s.io/cluster-api/pkg/util"
+	"strings"
+)
+
+// Deployer satisfies the ProviderDeployer(https://github.com/kubernetes-sigs/cluster-api/blob/master/cmd/clusterctl/clusterdeployer/clusterdeployer.go) interface.
+type Deployer struct {
+}
+
+// New returns a new Deployer.
+func New() *Deployer {
+	return &Deployer{}
+}
+
+// GetIP returns the IP of a machine, but this is going away.
+func (d *Deployer) GetIP(cluster *clusterv1.Cluster, machine *clusterv1.Machine) (string, error) {
+	if machine.ObjectMeta.Annotations != nil {
+		if ip, ok := machine.ObjectMeta.Annotations[constants.OpenstackIPAnnotationKey]; ok {
+			klog.Infof("Returning IP from machine annotation %s", ip)
+			return ip, nil
+		}
+	}
+
+	return "", errors.New("could not get IP")
+}
+
+// GetKubeConfig returns the kubeConfig after the bootstrap process is complete.
+func (d *Deployer) GetKubeConfig(cluster *clusterv1.Cluster, master *clusterv1.Machine) (string, error) {
+	ip, err := d.GetIP(cluster, master)
+	if err != nil {
+		return "", err
+	}
+
+	homeDir, ok := os.LookupEnv("HOME")
+	if !ok {
+		return "", fmt.Errorf("unable to use HOME environment variable to find SSH key: %v", err)
+	}
+
+	machineSpec, err := providerv1.MachineSpecFromProviderSpec(master.Spec.ProviderSpec)
+	if err != nil {
+		return "", err
+	}
+
+	result := strings.TrimSpace(util.ExecCommand(
+		"ssh", "-i", homeDir+"/.ssh/openstack_tmp",
+		"-o", "StrictHostKeyChecking no",
+		"-o", "UserKnownHostsFile /dev/null",
+		"-o", "BatchMode=yes",
+		fmt.Sprintf("%s@%s", machineSpec.SshUserName, ip),
+		"echo STARTFILE; sudo cat /etc/kubernetes/admin.conf"))
+	parts := strings.Split(result, "STARTFILE")
+	if len(parts) != 2 {
+		return "", nil
+	}
+	return strings.TrimSpace(parts[1]), nil
+}
